pyproject.toml
[project]
name = "streamlit-reto-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-identity>=1.23.0",
    "azure-storage-blob>=12.25.1",
    "dotenv>=0.9.9",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "streamlit>=1.45.1",
]
